+++ nvdla/sw/kmd/firmware/engine.c
@@ -97,7 +97,7 @@ dla_get_dma_cube_address(void *driver_context, void *task_data,
 					uint32_t destination)
 {
 	int32_t ret = 0;
-	uint64_t *pdst = (uint64_t *)dst_ptr;
+	uint32_t *pdst = (uint32_t *)dst_ptr;
        ret = dla_get_dma_address(driver_context, task_data, index,
 								dst_ptr, destination);
 	if (ret)
@@ -122,12 +122,12 @@ exit:
  */
 int
 dla_read_input_address(struct dla_data_cube *data,
-		       uint64_t *address,
+		       uint32_t *address,
 		       int16_t op_index,
 		       uint8_t roi_index,
 		       uint8_t bpp)
 {
-	uint64_t roi_desc_addr;
+	uint32_t roi_desc_addr;
 	int32_t ret = ERR(INVALID_INPUT);
 	struct dla_engine *en = dla_get_engine();
