+++ nvdla/sw/kmd/firmware/pdp.c
@@ -265,9 +265,9 @@ static int
 processor_pdp_program(struct dla_processor_group *group)
 {
 	int32_t ret = 0;
-	uint32_t reg, high, low;
-	uint64_t input_address = 0;
-	uint64_t output_address = 0;
+	uint16_t reg, high, low;
+	uint32_t input_address = 0;
+	uint32_t output_address = 0;
 	struct dla_engine *engine = dla_get_engine();
 	struct dla_pdp_op_desc *pdp_op;
 	struct dla_pdp_surface_desc *pdp_surface;
